Fixed training, new results

+++ TrianFlow/core/networks/structures/ransac.py
@@ -28,7 +28,7 @@ class reduced_ransac(nn.Module):
             for i in range(b):
                 nonzero_idx = torch.nonzero(mask[i,0,:]) # [nonzero_num,1]
                 if nonzero_idx.shape[0] == 0:
-                    select_idxs.append(torch.from_numpy(np.array([])).cuda()) #this shouldn't ever happen, but it did once
+                    select_idxs.append(torch.from_numpy(np.array([])).cuda().long()) #this shouldn't ever happen, but it did once
                     continue
                 rand_int = torch.randint(0, nonzero_idx.shape[0], [int(num)])
                 select_idx = nonzero_idx[rand_int, :] # [num, 1]
